+++ nodeconductor/core/authentication.py
@@ -26,9 +26,10 @@ class TokenAuthentication(rest_framework.authentication.TokenAuthentication):
         return auth
 
     def authenticate_credentials(self, key):
+        model = self.get_model()
         try:
-            token = self.model.objects.select_related('user').get(key=key)
-        except self.model.DoesNotExist:
+            token = model.objects.select_related('user').get(key=key)
+        except model.DoesNotExist:
             raise exceptions.AuthenticationFailed(_('Invalid token.'))
 
         if not token.user.is_active:
@@ -53,7 +54,13 @@ class TokenAuthentication(rest_framework.authentication.TokenAuthentication):
             msg = _('Invalid token. Token string should not contain spaces.')
             raise exceptions.AuthenticationFailed(msg)
 
-        return self.authenticate_credentials(auth[1])
+        try:
+            token = auth[1].decode()
+        except UnicodeError:
+            msg = _('Invalid token header. Token string should not contain invalid characters.')
+            raise exceptions.AuthenticationFailed(msg)
+
+        return self.authenticate_credentials(token)
 
 
 def user_capturing_auth(auth):